+++ test/brando/sites/redirects_test.exs
@@ -1,0 +1,7 @@
+defmodule Brando.Sites.RedirectsTest do
+  use ExUnit.Case
+
+  test "build regex" do
+    assert true
+  end
+end